Add cluster_ip module for retrieving container IP addresses
A new module, cluster_ip.py, has been added to the source code. This module provides a function, get_public_ips(cluster_name), that retrieves the public IP addresses of running containers in a specified ECS cluster. The function uses boto3 to interact with the ECS and EC2 clients to get the necessary information. Additionally, a main block has been included in the module to allow for standalone execution with the cluster name as a command line argument.